Modifications to allow model_url to be a local file path

+++ espaloma_charge/app.py
@@ -11,6 +11,8 @@ import torch
 from torch.utils.model_zoo import load_url
 import numpy as np
 from .utils import from_rdkit_mol
+
+# TODO: Do we really want to define this at file level, rather than within some kind of class?
 MODEL_URL = """
 https://github.com/choderalab/espaloma_charge/releases/download/v0.0.0/model.pt
 """
@@ -18,7 +20,7 @@ https://github.com/choderalab/espaloma_charge/releases/download/v0.0.0/model.pt
 def charge(
         molecule,
         total_charge: float = None,
-        model_url: str = MODEL_URL,
+        model_url: str = None,
     ) -> np.ndarray:
     """Assign machine-learned AM1-BCC partial charges to a molecule.
 
@@ -30,17 +32,28 @@ def charge(
     total_charge : float = 0.0
 
 
-    model_url : str
-        Model url.
+    model_url : str, optional, default=None
+        URL or filepath to retrieve the model from.
+        If None, the default MODEL_URL (defined at file level) is used
 
     Returns
     -------
     np.ndarray : (n_atoms, ) array of partial charges.
 
     """
+    if model_url is None:
+        model_url = MODEL_URL
+
+    # TODO: Can we memoize/cache the model so we don't have to retrieve it every time we invoke charging?
     with tempfile.TemporaryDirectory() as tempdir:
         target_path = os.path.join(tempdir, "model.pt")
-        request.urlretrieve(model_url, target_path)
+        if os.path.exists(model_url):
+            # model_url is a local filepath
+            import shutil
+            shutil.copyfile(model_url, target_path)
+        else:
+            # See if model_url is a URL
+            request.urlretrieve(model_url, target_path)
         model = torch.load(target_path, map_location="cpu")
     if total_charge is None:
         total_charge = Chem.GetFormalCharge(molecule)
